Time: 6 ms (14.74%), Space: 5.8 MB (96.18%) - LeetHub

+++ 50-powx-n/50-powx-n.cpp
@@ -17,7 +17,7 @@ public:
             }
         }
         if(n<0)
-            ans = (double)1.0/(double)(ans);
+            ans = (double)1/ans;
         return ans;
     }
 };
